pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iam-audit-lab"
version = "0.1.0"
description = "IAM audit tools for AWS, Azure, GCP, and Entra — identify excessive permissions, orphaned accounts, and MFA gaps"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
  { name = "Hiago Kin Levi" }
]
keywords = ["iam", "audit", "security", "aws", "azure", "gcp", "cloud", "permissions"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  # CLI framework
  "click>=8.1",

  # Data validation and settings
  "pydantic>=2.6",
  "python-dotenv>=1.0",

  # AWS
  "boto3>=1.34",

  # Azure and Entra
  "azure-identity>=1.16",
  "azure-mgmt-authorization>=4.0",

  # GCP
  "google-cloud-iam>=2.14",

  # Structured logging
  "structlog>=24.1",

  # Report formatting
  "tabulate>=0.9",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
  "boto3-stubs[iam]>=1.34",
  "moto[iam]>=5.0",   # AWS mock library for testing
]

[project.scripts]
k1n-iam-audit = "cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["providers", "analyzers", "schemas", "reports", "cli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = [
  "S101",   # allow assert statements in tests
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
